Module EzHex.Hex

type t = string

val compare : t -> t -> int

val equal : t -> t -> bool

val encode : string -> t

encode s converts a string to a lowercase hexadecimal notation

val encodeU : string -> t

encodeU s converts a string to an uppercase hexadecimal notation

val decode : t -> string

decode hex converts a string in hexadecimal notation into its corresponding decoded string. Can raise Invalid_argument.

val encode_bytes : bytes -> t

encode s converts a string to a lowercase hexadecimal notation

val encodeU_bytes : bytes -> t

encodeU s converts a string to an uppercase hexadecimal notation

val decode_bytes : t -> bytes

decode hex converts a string in hexadecimal notation into its corresponding decoded string. Can raise Invalid_argument.